// Font-end/ac_base/NextSolutionsBase_web/src/models/enum/column.types.ts
export enum ColumnTypes {
    VIEW = 'VIEW',
    INPUT_COUNTER = 'INPUT_COUNTER',
    DATE_PICKER = 'DATE_PICKER',
    MULTI_SELECT_AUTOCOMPLETE = 'MULTI_SELECT_AUTOCOMPLETE',
    CHECKBOX = 'CHECKBOX',
    BASE64 = 'BASE64',
    IMG = 'IMG',
    IMG_CONTENT = 'IMG_CONTENT',
    IMG_NO_SECURE = 'IMG_NO_SECURE',
    IMG_CUSTOMER = 'IMG_CUSTOMER',
    // HTML = 'HTML',
    // INNER_TEXT = 'INNER_TEXT',
    BUTTON = 'BUTTON',
    INPUT = 'INPUT',
    LINK = 'LINK',
    STYLE_CSS = 'STYLE_CSS'
}